+++ apps/meteor/app/lib/server/methods/saveSetting.ts
@@ -4,6 +4,7 @@ import { Settings } from '@rocket.chat/models';
 import { Match, check } from 'meteor/check';
 import { Meteor } from 'meteor/meteor';
 
+import { updateAuditedByUser } from '../../../../server/settings/lib/auditedSettingUpdates';
 import { twoFactorRequired } from '../../../2fa/server/twoFactorRequired';
 import { getSettingPermissionId } from '../../../authorization/lib';
 import { hasPermissionAsync, hasAllPermissionAsync } from '../../../authorization/server/functions/hasPermission';
@@ -18,7 +19,7 @@ declare module '@rocket.chat/ddp-client' {
 }
 
 Meteor.methods<ServerMethods>({
-	saveSetting: twoFactorRequired(async (_id, value, editor) => {
+	saveSetting: twoFactorRequired(async function (_id, value, editor) {
 		const uid = Meteor.userId();
 		if (!uid) {
 			throw new Meteor.Error('error-action-not-allowed', 'Editing settings is not allowed', {
@@ -66,7 +67,14 @@ Meteor.methods<ServerMethods>({
 				break;
 		}
 
-		(await Settings.updateValueAndEditorById(_id, value as SettingValue, editor)).modifiedCount &&
+		const auditSettingOperation = updateAuditedByUser({
+			_id: uid,
+			username: (await Meteor.userAsync())!.username!,
+			ip: this.connection?.clientAddress || '',
+			useragent: this.connection?.httpHeaders['user-agent'] || '',
+		});
+
+		(await auditSettingOperation(Settings.updateValueAndEditorById, _id, value as SettingValue, editor)).modifiedCount &&
 			setting &&
 			void notifyOnSettingChanged({ ...setting, editor, value: value as SettingValue });