+++ apps/meteor/app/livechat/server/lib/Helper.ts
@@ -97,7 +97,7 @@ export const createLivechatRoom = async (
 	const source = extraRoomInfo.source || roomInfo.source;
 
 	if (settings.get<string>('Livechat_Require_Contact_Verification') === 'always') {
-		await LivechatContacts.updateContactChannel({ visitorId: _id, source }, { verified: false });
+		await LivechatContacts.setChannelVerifiedStatus({ visitorId: _id, source }, false);
 	}
 
 	const contactId = await migrateVisitorIfMissingContact(_id, source);
@@ -282,8 +282,6 @@ export const createLivechatSubscription = async (
 		rid,
 		name,
 		fname: name,
-		lowerCaseName: name.toLowerCase(),
-		lowerCaseFName: name.toLowerCase(),
 		alert: true,
 		open: true,
 		unread: 1,