+++ apps/meteor/client/methods/index.ts
@@ -1,7 +1,0 @@
-import './hideRoom';
-import './openRoom';
-import './pinMessage';
-import './starMessage';
-import './toggleFavorite';
-import './unpinMessage';
-import './updateMessage';