+++ packages/instance-status/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @rocket.chat/instance-status
 
+## 0.1.10-rc.1
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/models@1.0.1-rc.1
+  </details>
+
+## 0.1.10-rc.0
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/models@1.0.1-rc.0
+  </details>
+
 ## 0.1.9
 
 ### Patch Changes